requirements.txt
pyqt5
pyqt5-tools
psycopg2
python-dotenv